Aligned: SharedUils SmartDI SharedForms
This library will be broken down into categories shortly, so is probably the last of this series:

(1) Navigation Guidance using an AppStateManager -- this will eliminate the current dependence on the SmarttDI for this file (SharedForms).
(2) FormsTplGuidance -- Will move all of the lifecycle aware stuff into that, and find ways to improve TPL compliance and get rid of the async/void dilemma wherever possible.
(3 App Scaling -- will be done in a small file to allow you to scale fonts, margins, etc., by the device resolution. In this way, the app will always appear to be correctly sizem, regardless of the device. Currently that code is hosted in FormsUtils.